refactor: IPC 통신 로직을 Infrastructure 레이어로 분리 (#2)
