+++ src/features/reminder/domain/reminder.ts
@@ -1,3 +1,5 @@
+import { Category } from '@src/shared/constants';
+
 /**
  * 단일 리마인더 할 일 항목의 구조
  */
@@ -20,3 +22,17 @@ export interface ReminderSectionData {
   isFixed: boolean; // 시스템 고정 섹션 여부 (삭제 불가)
   items: Reminder[];
 }
+
+const today = new Date();
+export const initialSections: ReminderSectionData[] = [
+  { id: Category.EVERYDAY, title: 'Everyday', isFixed: true, items: [
+    { id: 1, text: '약 먹기', time: new Date(new Date(today).setHours(14, 0, 0, 0)), isAllDay: false, notified: true, done: true },
+    { id: 2, text: '알고리즘 문제 풀기', time: new Date(new Date(today).setHours(16, 0, 0, 0)), isAllDay: false, notified: false, done: false },
+    { id: 3, text: '산책하기', time: new Date(new Date(today).setHours(23, 0, 0, 0)), isAllDay: false, notified: false, done: true },
+  ]},
+  { id: Category.TODO, title: 'To Do', isFixed: true, items: [
+    { id: 4, text: '책 반납하기', isAllDay: true, notified: false, done: false },
+    { id: 5, text: '편의점 택배 보내고 오기', time: new Date(new Date(today).setHours(16, 0, 0, 0)), isAllDay: false, notified: false, done: false },
+  ]},
+  { id: Category.WORK, title: 'Work', isFixed: false, items: [] },
+];
